代码优化，增加添加班级功能

+++ e-server/src/main/java/com/ecode/interceptor/JwtTokenAdminInterceptor.java
@@ -50,7 +50,7 @@ public class JwtTokenAdminInterceptor implements HandlerInterceptor {
         try {
             log.info("jwt校验:{}", token);
             Claims claims = JwtUtil.parseJWT(jwtProperties.getUserSecretKey(), token);
-            Long empId = Long.valueOf(claims.get(JwtClaimsConstant.USER_ID).toString());
+            Integer empId = Integer.valueOf(claims.get(JwtClaimsConstant.USER_ID).toString());
             log.info("当前用户id：", empId);
             BaseContext.setCurrentId(empId);
             //3、通过，放行
